+++ src/renderer/src/components/common-stock-select/index.ts
@@ -1,0 +1,1 @@
+export { CommonStockSelect } from './common-stock-select';